pyproject.toml
[project]
name = "python-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "faiss-cpu>=1.11.0.post1",
    "google-generativeai>=0.8.5",
    "ipython>=9.4.0",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-tavily>=0.2.11",
    "langchain-text-splitters>=0.3.9",
    "langchain[openai]>=0.3.27",
    "langgraph>=0.6.4",
    "pypdf>=5.9.0",
]
